requirements.txt
streamlit
pdfplumber
python-docx
nltk
openai==0.28
python-dotenv
reportlab
